Add Eq and Show as constraints on Resolvable.
We need Eq to resolve some conflicts.

We somewhat reluctantly want Show so that if an unexpected conflict occurs, we
have a hope of being able to use the debug machinery to tell what's going on.

As an example case of unexpected conflict, consider serialising, then
deserialising, a Double, and hoping that each value still compares as equal to
the other (hint: they won't).